refactor: logout 로직 annotation Controller로 동작하도록 변경
