--- app/src/main/java/com/techcourse/controller/manual/LogoutController.java
@@ -1,16 +1,0 @@
-package com.techcourse.controller.manual;
-
-import com.techcourse.controller.UserSession;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
-
-public class LogoutController implements Controller {
-
-    @Override
-    public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        final var session = req.getSession();
-        session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
-    }
-}
